+++ src/threaded_writer.c
@@ -137,10 +137,10 @@ int32_t jls_twr_run(struct jls_twr_s * self) {
                     break;
                 case MSG_ANNOTATION:
                     rc = jls_wr_annotation(self->wr, hdr.h.annotation.signal_id, hdr.h.annotation.timestamp,
+                                           hdr.h.annotation.y,
                                            hdr.h.annotation.annotation_type,
                                            hdr.h.annotation.group_id,
                                            hdr.h.annotation.storage_type,
-                                           hdr.h.annotation.y,
                                            (const uint8_t *) payload, payload_sz);
                     break;
                 case MSG_UTC:
@@ -293,10 +293,10 @@ int32_t jls_twr_fsr_f32(struct jls_twr_s * self, uint16_t signal_id,
 }
 
 int32_t jls_twr_annotation(struct jls_twr_s * self, uint16_t signal_id, int64_t timestamp,
+                           float y,
                            enum jls_annotation_type_e annotation_type,
                            uint8_t group_id,
                            enum jls_storage_type_e storage_type,
-                           float y,
                            const uint8_t * data, uint32_t data_size) {
     struct msg_header_s hdr = {
             .msg_type = MSG_ANNOTATION,